requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bootstrap4==0.1.0
dj-static==0.0.6
Django==4.1.7
django-bootstrap4==22.3
django-stdimage==6.0.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
soupsieve==2.4
sqlparse==0.4.3
static3==0.7.0
